requirements.txt
dwave-ocean-sdk>=4.3.0
dwave-neal

notebook~=7.0
jupyter

pandas
matplotlib~=3.0
seaborn

